+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/DoubleEditorFactory.java
@@ -19,7 +19,10 @@ import net.sourceforge.pmd.properties.PropertySource;
 
 /**
  * @author Brian Remedios
+ * @deprecated This editor factory will be removed without replacement. This was only used for supporting the UI
+ *             of the plugin and is considered internal API now.
  */
+@Deprecated // for removal
 public final class DoubleEditorFactory extends AbstractRealNumberEditor<Double> {
 
     public static final DoubleEditorFactory INSTANCE = new DoubleEditorFactory();
@@ -59,8 +62,6 @@ public final class DoubleEditorFactory extends AbstractRealNumberEditor<Double> 
 
                 source.setProperty(desc, newValue);
                 listener.changed(source, desc, newValue);
-
-                adjustRendering(source, desc, spinner);
             }
         });