+++ net.sourceforge.pmd.eclipse.plugin/src/main/java/net/sourceforge/pmd/eclipse/ui/preferences/editors/SWTUtil.java
@@ -14,7 +14,6 @@ import java.util.Set;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Combo;
-import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Event;
 import org.eclipse.swt.widgets.Listener;
@@ -24,20 +23,28 @@ import net.sourceforge.pmd.eclipse.plugin.PMDPlugin;
 /**
  *
  * @author Brian Remedios
+ * @deprecated This is internal API and will be removed.
  */
+@Deprecated // for removal
 public final class SWTUtil {
 
     private static PMDPlugin plugin = PMDPlugin.getDefault();
 
-    private static final String TOOLTIP_SUFFIX = ".tooltip";
-
     private SWTUtil() {
     }
 
+    /**
+     * @deprecated use {@link PMDPlugin#logInformation(String)} instead.
+     */
+    @Deprecated // for removal
     public static void logInfo(String message) {
         plugin.logInformation(message);
     }
 
+    /**
+     * @deprecated use {@link PMDPlugin#logError(org.eclipse.core.runtime.IStatus)} instead.
+     */
+    @Deprecated // for removal
     public static void logError(String message, Throwable error) {
         plugin.logError(message, error);
     }
@@ -46,7 +53,9 @@ public final class SWTUtil {
      * Let the buttons operate as a radio group, with only one button selected at a time.
      * 
      * @param buttons
+     * @deprecated this is not used and will be removed
      */
+    @Deprecated // for removal
     public static void asRadioButtons(final Collection<Button> buttons) {
 
         Listener listener = new Listener() {
@@ -66,13 +75,19 @@ public final class SWTUtil {
         }
     }
 
-    // TODO move this to to Collections utility
+    /**
+     * @deprecated this is not used and will be removed
+     */
+    @Deprecated // for removal
     public static Set<String> asStringSet(String input, char separator) {
         List<String> values = Arrays.asList(input.split("" + separator));
         return new HashSet<>(values);
     }
 
-    // TODO move this to to Collections utility
+    /**
+     * @deprecated this is not used and will be removed
+     */
+    @Deprecated // for removal
     public static String asString(Collection<String> values, char separator) {
         if (values == null || values.isEmpty()) {
             return "";
@@ -87,44 +102,58 @@ public final class SWTUtil {
         return sb.toString();
     }
 
-    public static void setEnabled(Control control, boolean flag) {
-        if (control == null || control.isDisposed()) {
-            return;
-        }
-        control.setEnabled(flag);
-    }
-
+    /**
+     * @deprecated this is not used and will be removed
+     */
+    @Deprecated // for removal
     public static void setEnabled(Control[] controls, boolean state) {
-        for (Control control : controls) {
-            setEnabled(control, state);
-        }
+        net.sourceforge.pmd.eclipse.util.internal.SWTUtil.setEnabled(Arrays.asList(controls), state);
     }
 
-    public static void setEnabled(Collection<Control> controls, boolean flag) {
-        for (Control control : controls) {
-            setEnabled(control, flag);
-        }
-    }
-
+    /**
+     * @deprecated This is internal API and will be removed.
+     */
+    @Deprecated // for removal
     public static void setEnabledRecursive(Control[] controls, boolean state) {
-        for (Control control : controls) {
-            if (control instanceof Composite) {
-                setEnabledRecursive(((Composite) control).getChildren(), state);
-            }
-            setEnabled(control, state);
-        }
+        net.sourceforge.pmd.eclipse.util.internal.SWTUtil.setEnabledRecursive(controls, state);
     }
 
+    /**
+     * @deprecated This is internal API and will be removed.
+     */
+    @Deprecated // for removal
+    public static void setEnabled(Control control, boolean flag) {
+        net.sourceforge.pmd.eclipse.util.internal.SWTUtil.setEnabled(control, flag);
+    }
+
+    /**
+     * @deprecated This is internal API and will be removed.
+     */
+    @Deprecated // for removal
+    public static void setEnabled(Collection<Control> controls, boolean flag) {
+        net.sourceforge.pmd.eclipse.util.internal.SWTUtil.setEnabled(controls, flag);
+    }
+
+    /**
+     * @deprecated Use {@link PMDPlugin#getStringTable()} instead.
+     */
+    @Deprecated // for removal
     public static String stringFor(String key) {
-        return plugin.getStringTable().getString(key);
+        return net.sourceforge.pmd.eclipse.util.internal.SWTUtil.stringFor(key);
     }
 
+    /**
+     * @deprecated Use {@link PMDPlugin#getStringTable()} with suffix {@code .tooltip} instead.
+     */
+    @Deprecated // for removal
     public static String tooltipFor(String key) {
-        String ttKey = key + TOOLTIP_SUFFIX;
-        String tooltip = stringFor(ttKey);
-        return ttKey.equals(tooltip) ? stringFor(key) : tooltip;
+        return net.sourceforge.pmd.eclipse.util.internal.SWTUtil.tooltipFor(key);
     }
 
+    /**
+     * @deprecated this is not used and will be removed
+     */
+    @Deprecated // for removal
     public static void releaseListeners(Control control, int listenerType) {
         Listener[] listeners = control.getListeners(listenerType);
         for (Listener listener : listeners) {
@@ -132,31 +161,27 @@ public final class SWTUtil {
         }
     }
 
+    /**
+     * @deprecated This is internal API and will be removed.
+     */
+    @Deprecated // for removal
     public static String[] labelsIn(Object[][] items, int columnIndex) {
-
-        String[] labels = new String[items.length];
-        for (int i = 0; i < labels.length; i++) {
-            labels[i] = items[i][columnIndex].toString();
-        }
-        return labels;
+        return net.sourceforge.pmd.eclipse.util.internal.SWTUtil.labelsIn(items, columnIndex);
     }
 
-    public static void deselectAll(Combo combo) {
-        int count = combo.getItems().length;
-        for (int i = 0; i < count; i++) {
-            combo.deselect(i);
-        }
-    }
-
+    /**
+     * @deprecated This is internal API and will be removed.
+     */
+    @Deprecated // for removal
     public static String[] i18lLabelsIn(Object[][] items, int columnIndex) {
+        return net.sourceforge.pmd.eclipse.util.internal.SWTUtil.i18lLabelsIn(items, columnIndex);
+    }
 
-        String[] labels = labelsIn(items, columnIndex);
-        String xlation;
-
-        for (int i = 0; i < labels.length; i++) {
-            xlation = stringFor(labels[i]);
-            labels[i] = xlation == null ? labels[i] : xlation;
-        }
-        return labels;
+    /**
+     * @deprecated This is internal API and will be removed.
+     */
+    @Deprecated // for removal
+    public static void deselectAll(Combo combo) {
+        net.sourceforge.pmd.eclipse.util.internal.SWTUtil.deselectAll(combo);
     }
 }